moved store interface for elasticsearch into its own file
